refactor(acp): extract useAcpModelInfo and useAgentModesForBackend hooks
